requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2

# AI
anthropic==0.34.2

# Scheduling
apscheduler==3.10.4

# Integrations
notion-client==2.2.1
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
icalendar==6.0.0
caldav==1.3.9

# Notifications
plyer==2.1.0
httpx==0.27.2

# CLI / UX
rich==13.8.1
typer==0.12.5
python-dotenv==1.0.1

# Utilities
pytz==2024.2
python-dateutil==2.9.0
croniter==3.0.3
